Adding exceptions for fetching forecast and mapping wind, modifying mapper, fetch service, changing String date to creationDate and forecastDate in Instant format and windSpeed to float

+++ src/main/java/com/sda/weather/forecast/ForecastOpenWeatherResponse.java
@@ -21,7 +21,7 @@ public class ForecastOpenWeatherResponse {
 
     @Data
     public static class Wind {
-        private int speed;
+        private float speed;
         @JsonProperty("deg")
         private float degree;
     }
